Drink, Food, Snack fully implemented.
Altered the toString method of Snack ever so slightly - removed the space between the newline character and first character.

Relocated the main methods to the bottom of the classes. Altered the toString methods in Food and Drink to include extra attributes introduced by the extension.

Implemented the validateSugarContent method and correctly implemented the calculatePrice method for the Drink class.